INTERMEDIATE COMMIT
Major in-progress refactor.  Redo the installation process to be
done via a recipe.  This will only draw on attributes for defaults.
This means that Redis installation settings will be standardized across
an organization with the exception of Redis version. Version is special
because we will not install the version specified if a version already
exists.

Setting up a redis server is now done via a LWRP block.  By default all
node attributes will be used however anything set in the LWRP block will
override these settings.